+++ R/frag.simul.process.R
@@ -142,7 +142,7 @@
   # default behaviour:
   v.to.disturb <- sample(seq(1, igraph::gorder(g)), nr.v.to.disturb)
   # if asymmetric.transport.from is set:
-  if( ! is.null(asymmetric.transport.from)){
+  if(asymmetric.transport.from != 0){
     if(nr.v.to.disturb <= length(igraph::V(g)[ igraph::V(g)$layer == asymmetric.transport.from]) ){
       v.to.disturb <- sample(igraph::V(g)[ igraph::V(g)$layer == asymmetric.transport.from], nr.v.to.disturb)
     } else{
@@ -184,6 +184,7 @@ frag.simul.process <- function(initial.layers=2, n.components=NULL, vertices=Inf
     if(missing(components.balance)) components.balance <- params$components.balance
     if(missing(disturbance)) disturbance <- params$disturbance
     if(missing(aggreg.factor)) aggreg.factor <- params$aggreg.factor
+    if(missing(asymmetric.transport.from)) asymmetric.transport.from <- 0
     if(missing(planar))  planar <- params$planar
     if(is.na(planar)) {
       planar <- FALSE
@@ -230,8 +231,8 @@ frag.simul.process <- function(initial.layers=2, n.components=NULL, vertices=Inf
   }
 
   if( ! is.null(asymmetric.transport.from) ){
-    if(! asymmetric.transport.from %in% c(1, 2, "1", "2")){
-      stop("The 'asymmetric.transport.from' parameter must have a value in 1 or 2.")
+    if(! asymmetric.transport.from %in% c(0, 1, 2, "0" ,"1", "2")){
+      stop("Values for 'asymmetric.transport.from' must be one of 1, 2 or 0.")
     }
   }